Updated for new double standard rather than floats.

+++ program/include/camera.hpp
@@ -1,18 +1,18 @@
 #pragma once
 
 #include "cse/camera.hpp"
-#include "glm/ext/vector_float3.hpp"
+#include "glm/ext/vector_double3.hpp"
 
 namespace csg
 {
   class camera : public cse::camera
   {
   public:
-    camera(const glm::vec3 &translation_);
+    camera(const glm::dvec3 &translation_);
 
   private:
-    float max_velocity{200.0f};
-    float friction{50.0f};
+    double max_velocity{200.0};
+    double friction{50.0};
     double fov_change{10.0};
   };
 }
